Create Non Repeating Character.java
Non Repeating Character
Given a string S consisting of lowercase Latin Letters. Find the first non-repeating character in S.

Example 1:

Input: S = hello Output: h Explanation: In the given string, the first character which is non-repeating is h, as it appears first and there is no other h in the string.

Example 2:

Input: S = zxvczbtxyzvy Output: c Explanation: In the given string, c is the character which is non-repeating.

Your Task:
You need to take string S as a parameter and output the character. If there is no non-repeating character then return -1 .

Constraints:
1 <= N <= 103