LCA in BST
LCA In A BST
You are given a binary search tree of integers with N nodes. You are also given references to two nodes P and Q from this BST. Your task is to find the lowest common ancestor(LCA) of these two given nodes. The lowest common ancestor for two nodes P and Q is defined as the lowest node that has both P and Q as descendants (where we allow a node to be a descendant of itself) A binary search tree (BST) is a binary tree data structure which has the following properties. • The left subtree of a node contains only nodes with data less than the node’s data. • The right subtree of a node contains only nodes with data greater than the node’s data. • Both the left and right subtrees must also be binary search trees.

Input Format: The first line contains an integer 'T' which denotes the number of test cases or queries to be run. Then the test cases follow.

The first line of each test case contains two space separated integers P and Q, the nodes whose LCA we have to find.

The second line of each test case contains the elements of the BST in the level order form separated by a single space.

If any node does not have a left or right child, take -1 in its place. Refer to the example below.

Output Format : For each test case print the LCA of nodes P and Q.

Output for every test case will be printed in a separate line.

Constraint : 1 <= T <= 100 1 <= N <= 5000 0 <= Data <= 10^6 and Data != -1

Sample Input 1 : 2 3 5 2 1 3 -1 -1 -1 5 -1 -1 1 3 2 1 4 -1 -1 3 -1 -1 -1

Sample Output 1 : 3 2

Sample Input 2: 1 1 1 3 2 -1 1 -1 -1 -1

Sample Output 2: 1